Setup connection to the service worker
The content script and the service worker need to exchange messages via
chrome ports. This should also allow integration with the API.

This PR refactors the code and formelizes the several connections in the
extension. It also adds `proxy` method for the content script to forward
messages between the inpage script and the service worker.