Add sample page to start a KYC process
On the attester page, we need a way to figure out if we should start a
KYC process or not.

The way we discussed doing so is by checking if the user has the
necessary fields for a type of verification. The necessary fields are
defined by the Attester, so the API must receive a list of fields to
check if the user has them defined.

For now it's ok to do this operation without the user having to give
permissions since we're not actually accessing private information.